add filter_data_context.rs
