+++ src/engine.rs
@@ -5,8 +5,8 @@ use crate::cosmetic_filter_cache::{CosmeticFilterCache, UrlSpecificResources};
 use crate::cosmetic_filter_cache_builder::CosmeticFilterCacheBuilder;
 use crate::filters::cosmetic::CosmeticFilter;
 use crate::filters::fb_builder::EngineFlatBuilder;
-use crate::filters::fb_network::{FilterDataContext, FilterDataContextRef};
 use crate::filters::fb_network_builder::NetworkRulesBuilder;
+use crate::filters::filter_data_context::{FilterDataContext, FilterDataContextRef};
 use crate::filters::network::NetworkFilter;
 use crate::flatbuffers::containers::flat_serialize::FlatSerialize;
 use crate::flatbuffers::unsafe_tools::VerifiedFlatbufferMemory;
