+++ src/filters/fb_network.rs
@@ -1,13 +1,11 @@
 //! Flatbuffer-compatible versions of [NetworkFilter] and related functionality.
 
-use std::collections::HashMap;
-
+use crate::filters::filter_data_context::FilterDataContext;
 use crate::filters::network::{NetworkFilterMask, NetworkFilterMaskHelper, NetworkMatchable};
-use crate::flatbuffers::unsafe_tools::{fb_vector_to_slice, VerifiedFlatbufferMemory};
+use crate::flatbuffers::unsafe_tools::fb_vector_to_slice;
 
 use crate::regex_manager::RegexManager;
 use crate::request::Request;
-use crate::utils::Hash;
 
 #[allow(unknown_lints)]
 #[allow(
@@ -71,34 +69,6 @@ impl ExactSizeIterator for FlatPatternsIterator<'_> {
     #[inline(always)]
     fn len(&self) -> usize {
         self.len
-    }
-}
-
-#[cfg(feature = "single-thread")]
-pub(crate) type FilterDataContextRef = std::rc::Rc<FilterDataContext>;
-#[cfg(not(feature = "single-thread"))]
-pub(crate) type FilterDataContextRef = std::sync::Arc<FilterDataContext>;
-
-// The struct is used to store the flatbuffer and supporting data
-// for both network filter and cosmetic filters.
-// Supposed to be stored via FilterDataContextRef to avoid copying the data.
-pub(crate) struct FilterDataContext {
-    pub(crate) memory: VerifiedFlatbufferMemory,
-    pub(crate) unique_domains_hashes_map: HashMap<Hash, u32>,
-}
-
-impl FilterDataContext {
-    pub(crate) fn new(memory: VerifiedFlatbufferMemory) -> FilterDataContextRef {
-        // Reconstruct the unique_domains_hashes_map from the flatbuffer data
-        let root = memory.root();
-        let mut unique_domains_hashes_map: HashMap<crate::utils::Hash, u32> = HashMap::new();
-        for (index, hash) in root.unique_domains_hashes().iter().enumerate() {
-            unique_domains_hashes_map.insert(hash, index as u32);
-        }
-        FilterDataContextRef::new(Self {
-            memory,
-            unique_domains_hashes_map,
-        })
     }
 }
