+++ src/filters/mod.rs
@@ -7,4 +7,5 @@ pub mod cosmetic;
 pub(crate) mod fb_builder;
 pub(crate) mod fb_network;
 pub(crate) mod fb_network_builder;
+pub(crate) mod filter_data_context;
 pub mod network;
